backend/requirements/base.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
openai>=1.0.0,<2.0.0
httpx>=0.28.0,<1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
